wip(today's fortune): Cynic, n.: 	Experienced.
