+++ sources/affinity/model/README.md
@@ -5,6 +5,6 @@ You can find the latest version of the spec [here](https://developer.affinity.co
 ## How to regenerate
 
 1. Download the new spec from the link above.
-2. Put it into this folder (don't forget to remove the old version)
+2. Put it into this folder.
 3. Run `pip install -r requirements.txt` or any other means to get [datamodel-code-generator](https://pypi.org/project/datamodel-code-generator/).
 4. Run `./generate_model.sh`
