+++ sources/affinity/model/v2_spec.json
